auth_service/requirements.txt
Flask>=2.0
requests>=2.20
PyJWT>=2.0
gunicorn>=20.0